langchain/core-components/1-agents/1-static-model.py
"""静态模型示例：固定使用一个 LLM 配置创建 agent。"""

import os

from langchain.agents import create_agent
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv
from openai import api_key
from pydantic import SecretStr

load_dotenv(dotenv_path=".env")

# 从环境变量读取 SiliconFlow 相关配置。
api_key = os.getenv("SILICONFLOW_API_KEY")
base_url = os.getenv("SILICONFLOW_BASE_URL")
temperature = os.getenv("SILICONFLOW_TEMPERATURE", 0.2)

if not api_key or not base_url:
    raise ValueError(
        "SILICONFLOW_API_KEY and SILICONFLOW_BASE_URL must be set in the .env file"
    )

model = ChatOpenAI(
    # 这里的模型是固定值，不会根据上下文动态切换。
    model="Pro/MiniMaxAI/MiniMax-M2.5",
    api_key=SecretStr(api_key),
    base_url=base_url,
    temperature=float(temperature),
    timeout=60,
)
