pyproject.toml
[project]
name = "langchain-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.4.11",
    "fastapi>=0.135.1",
    "jupyter>=1.1.1",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "langsmith>=0.7.20",
    "loguru>=0.7.3",
    "notebook>=7.5.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
